requirements.txt
selenium
bs4
pandas
openpyxl
pymongo
python-dotenv
openai